backend/requirements.txt
fastapi
uvicorn
ortools
pydantic
pytest
sqlite3-api
chromadb
requests
google-genai
google-api-python-client
google-auth-oauthlib
python-dotenv
playwright
beautifulsoup4
duckduckgo-search
moviepy
cdp-sdk
psutil
yfinance
youtube-transcript-api
bandit
stripe
